refactor(cli): extract walk + filter pipeline to crate::walk
`lint`, `clean`, and `rename` each carried byte-for-byte identical
copies of `collect_entries` and `entry_matches_filter` — scanner +
ignore rules + `EntryKind::File` filter + project-relative path
matching. Move them to a single `crate::walk::collect_entries` so the
three subcommands stop drifting on filter semantics or error
messages, and so a future `progest search` (M3) inherits the same
walker for free.

Net: -122 +76 lines (one new module, three duplicate fn pairs
deleted).
